requirements.txt
# local package
-e .

# external requirements
torch
torchvision
scikit-learn
neurokit2==0.1.7
matplotlib
pandas
numpy
skorch
tqdm
click
python-dotenv>=0.5.1
split-folder
